move OpaqueHandle to new file
