+++ src/os/threading.hpp
@@ -4,21 +4,6 @@
 #pragma once
 #include "foundation/foundation.hpp"
 
-template <usize k_size>
-struct OpaqueHandle {
-    template <typename T>
-    T& As() {
-        static_assert(sizeof(T) == k_size);
-        static_assert(alignof(T) <= __alignof__(data));
-        return *reinterpret_cast<T*>(data);
-    }
-    template <typename T>
-    T const& As() const {
-        return const_cast<OpaqueHandle*>(this)->As<T>();
-    }
-    alignas(8) u8 data[k_size];
-};
-
 struct NativeHandleSizes {
     usize thread;
     usize mutex;
